+++ src/components/FoodModal.js
@@ -54,7 +54,7 @@ const ModalStyles = styled(CardStyles)`
 
 export default function FoodModal({ handleClick, foodId, returnData }) {
   // path to the fda api
-  const url = `https://api.nal.usda.gov/fdc/v1/food/${foodId}?api_key=${process.env.REACT_APP_API_KEY}`;
+  const url = `https://api.nal.usda.gov/fdc/v1/food/${foodId}?api_key=${process.env.REACT_APP_API_KEY}&nutrients=208`;
   // fetch data
   const { loading, data, error } = useFetch(url);
   const [foodNutrients, setFoodNutrients] = useState(null);
@@ -152,11 +152,11 @@ export default function FoodModal({ handleClick, foodId, returnData }) {
                 }
                 defaultValue={foodNutrients.portion}
               >
-                {/* {foodNutrients.portions.map((el, i) => (
+                {foodNutrients.portions.map((el, i) => (
                   <option key={el.id} value={el.gramWeight}>
                     {el.portionDescription}
                   </option>
-                ))} */}
+                ))}
               </select>
             </>
           ) : (
